main

+++ FahToCelsius.cpp
@@ -2,11 +2,11 @@
 using namespace std;
 
 int main() {
-	int fahrenheit;
+	float fahrenheit;
     cout << "Enter temperature in Fahrenheit: ";
     cin >> fahrenheit;
 
-    float celsius = (fahrenheit - 32) * 5.0 / 9.0;
+    float celsius = (fahrenheit - 32.0) * 5.0 / 9.0;
     cout << "Temperature in Celsius: " << celsius << endl;
 
     return 0;
